refactored functions from Landing Page to app.js
